Add files via upload
We found that there are seasures in different times of the day, some days like Friday, Saturday and Tuesday are having more seasures and months like July and in March seasures are most likely to ocuur. To define if there is a relationships between seasures and time of the day or lenght over time of the day or type of a seasure over time of the day we need more data.